Make everything work and accord to 1 style (except LQ)
