--- methods/i/constants.py
@@ -1,75 +1,0 @@
-A1 = [
-    [26, 65, -32, -35, 91, 70, 27, -96, 25],
-    [-55, 88, -11, 71, -15, -18, -10, 29, -46],
-    [-69, 12, -78, 52, -93, -77, -95, 3, -20],
-    [-86, -11, -60, -83, -1, -39, 54, 13, 41],
-    [-61, -22, 99, -56, -64, -79, -46, 53, -58],
-    [41, -46, -18, 84, 69, 38, -71, -84, -26],
-    [87, -14, -60, 40, 12, 13, -58, -18, -50],
-    [93, -91, 65, -85, -26, -12, 91, 4, 58],
-    [33, -34, -75, -72, -66, 15, 84, 11, -72],
-]
-
-b1 = [
-    [-69],
-    [1],
-    [-45],
-    [-79],
-    [-97],
-    [-72],
-    [87],
-    [44],
-    [-15],
-]
-
-A2 = [
-    [188, -32, -9, 7, -61, 49, -53, 23, -31, 62],
-    [-32, 101, 11, -32, -80, 1, 53, -7, 94, -109],
-    [-9, 11, 231, 96, -62, -12, 72, -40, 72, -77],
-    [7, -32, 96, 174, -37, -119, 19, 7, -27, 27],
-    [-61, -80, -62, -37, 184, -34, -115, 45, -53, 46],
-    [49, 1, -12, -119, -34, 182, 80, -56, -4, 18],
-    [-53, 53, 72, 19, -115, 80, 231, -101, 70, -50],
-    [23, -7, -40, 7, 45, -56, -101, 227, -10, 23],
-    [-31, 94, 72, -27, -53, -4, 70, -10, 241, -155],
-    [62, -109, -77, 27, 46, 18, -50, 23, -155, 212],
-]
-
-b2 = [
-    [-44],
-    [74],
-    [87],
-    [247],
-    [-39],
-    [168],
-    [192],
-    [-199],
-    [-194],
-    [243],
-]
-
-A3 = [
-    [390, 201, 0, 0, 0, 0, 0, 0, 0, 0],
-    [21, -135, -53, 0, 0, 0, 0, 0, 0, 0],
-    [0, -271, 868, -203, 0, 0, 0, 0, 0, 0],
-    [0, 0, -7, 596, 160, 0, 0, 0, 0, 0],
-    [0, 0, 0, -47, -208, -96, 0, 0, 0, 0],
-    [0, 0, 0, 0, -24, 347, -17, 0, 0, 0],
-    [0, 0, 0, 0, 0, -96, 220, -40, 0, 0],
-    [0, 0, 0, 0, 0, 0, 107, 546, -204, 0],
-    [0, 0, 0, 0, 0, 0, 0, -301, -740, -364],
-    [0, 0, 0, 0, 0, 0, 0, 0, 48, -154],
-]
-
-b3 = [
-    [446],
-    [62],
-    [-783],
-    [264],
-    [-747],
-    [-732],
-    [-803],
-    [-716],
-    [-664],
-    [-608],
-]
