--- methods/max/var30_constants.py
@@ -1,75 +1,0 @@
-A1 = [
-    [-27, -93, -74, -45, 67, 74, 17, 96, -36],
-    [59, 79, -66, 45, -90, 20, -58, -61, -62],
-    [73, -82, -1, 9, -79, -27, 40, 26, 52],
-    [-39, 91, -95, 80, 52, -26, -14, -10, -42],
-    [-23, -64, -10, 61, -33, -98, -28, 12, 22],
-    [17, 45, 23, 58, 27, -27, -78, 99, -64],
-    [-39, 27, 6, -66, 87, 32, -15, -36, 82],
-    [89, 40, -17, -44, -29, -31, 55, -52, 3],
-    [81, -60, 86, 42, 43, 29, 95, -70, -26],
-]
-
-B1 = [
-    [43],
-    [84],
-    [-7],
-    [38],
-    [59],
-    [93],
-    [-5],
-    [73],
-    [-10],
-]
-
-A2 = [
-    [688, -245, -382, 56, 166, -290, -5, -49, -347, -279],
-    [-245, 652, 382, 235, -226, -28, -73, -132, -148, 96],
-    [-382, 382, 705, 96, -41, 43, -30, 198, 182, 193],
-    [56, 235, 96, 611, 43, 280, -289 - 158, 74, 179],
-    [166, -226, -41, 43, 166, -4, -51, 143, 77, 66],
-    [-290, -28, 43, 280, -4, 677, 107, -151, 395, 183],
-    [-5, -73, -30, -289, -51, 107, 691, -158, 128, -74],
-    [-49, -132, 198, -158, 143, -151, -158, 498, 36, 223],
-    [-347, -148, 182, 74, 77, 395, 128, 36, 668, 299],
-    [-279, 96, 193, 179, 66, 183, -74, 223, 299, 751],
-]
-
-B2 = [
-    [-324],
-    [79],
-    [671],
-    [553],
-    [-120],
-    [-68],
-    [-173],
-    [15],
-    [644],
-    [22]
-]
-
-A3 = [
-    [949, 134, 0, 0, 0, 0, 0, 0, 0, 0],
-    [53, 994, 15, 0, 0, 0, 0, 0, 0, 0],
-    [0, -24, -339, -3, 0, 0, 0, 0, 0, 0],
-    [0, 0, 332 - 858, -373, 0, 0, 0, 0, 0],
-    [0, 0, 0, 41, -128, -13, 0, 0, 0, 0],
-    [0, 0, 0, 0, 72, 227, 88, 0, 0, 0],
-    [0, 0, 0, 0, 0, -267, 863, -208, 0, 0],
-    [0, 0, 0, 0, 0, 0, 74, 796, 3, 0],
-    [0, 0, 0, 0, 0, 0, 0, 72, 226, 7],
-    [0, 0, 0, 0, 0, 0, 0, 0, -55, -596],
-]
-
-B3 = [
-    [-144],
-    [933],
-    [240],
-    [391],
-    [441],
-    [-306],
-    [34],
-    [113],
-    [-687],
-    [124],
-]
