+++ python/cholesky.py
@@ -1,21 +1,24 @@
 from __future__ import print_function
+import time
+
 from math import sqrt
+from matrix_tool import zeros, print_vector, print_matrix, append_column, multiply
 import gauss_klassisch as gk
-from check import check
+from check import check_roots
 
 
 def solve(A, b):
   n = len(A)
   U = getU(A)
-  gk.print_matrix(U)
+  print_matrix(U)
   Ut = transpose(U)
 
   y = []
-  Utb = gk.append_column(Ut, b)
+  Utb = append_column(Ut, b)
   for row in xrange(n):
     y.append(gk.backward_left(Utb, row))
   print(y)
-  Uy = gk.append_column(U, [[e] for e in y])
+  Uy = append_column(U, [[e] for e in y])
   x = []
   for row in xrange(n - 1, -1, -1):
     x.append(gk.backward_right(Uy, row))
@@ -31,34 +34,33 @@ def determinant(U):
   return det**2
 
 
-def inverse(A, U):
+def inverse_triangle_right(U):
   # T = U^-1
   n = len(U)
   T = zeros(n, n)
   for i in xrange(n):
     T[i][i] = 1. / U[i][i]
-    for j in xrange(i+1, n):
+    for j in xrange(i + 1, n):
       for k in xrange(j):
         T[i][j] -= T[i][k] * U[k][j]
       T[i][j] /= U[j][j]
+  return T
 
+
+def inverse(A, U):
+  n = len(U)
+  T = inverse_triangle_right(U)
+  print
+  print("U^-1:")
+  print_matrix(T)
+  print("Check U^-1:")
+  print_matrix(multiply(T, U))
   B = zeros(n, n)
   for i in xrange(n):
     for j in xrange(n):
       for k in xrange(min(i, j), n):
         B[i][j] += T[i][k] * T[j][k]
   return B
-
-
-def zeros(n, m):
-  return [[0.] * m for _ in xrange(n)]
-
-
-def ones(n):
-  res = zeros(n, n)
-  for i in xrange(len(res)):
-    res[i][i] = 1
-  return res
 
 
 def getU(A):
@@ -91,25 +93,32 @@ def transpose(U):
   return Ut
 
 
-def mul(A, B):
-  n, m, k = len(A), len(A[0]), len(B[0])
-  C = zeros(n, k)
-  for i in xrange(n):
-    for j in xrange(k):
-      for d in xrange(m):
-        C[i][j] += A[i][d] * B[d][j]
-  return C
-
-
 def main():
-  from constants import A2, b2
-  x, U = solve(A2, b2)
-  print("Roots: {}".format(x))
-  check(A2, x, b2)
+  from constants import A2_21 as A, b2_21 as b
+  start_time = time.time()
+  x, U = solve(A, b)
+  solve_time = time.time()
+  print("Roots:")
+  print_vector(x)
+  check_roots(A, x, b)
   det = determinant(U)
+  det_time = time.time()
   print()
   print("Determinant: {}".format(det))
   print()
-  inversed = inverse(A2, U)
+  inversed = inverse(A, U)
+  inverse_time = time.time()
   print("Inverse matrix:")
-  gk.print_matrix_precise(inversed)
+  print_matrix(inversed, precision=8)
+  print("Check inverse matrix:")
+  print_matrix(multiply(inversed, A))
+
+  print("Time elapsed: ")
+  print(
+      "\tSolution:\t{} ms\n\tDeterminant:\t{} ms\n\tInversion:\t{} ms\n\tOverall:\t{} ms".
+      format((solve_time - start_time) * 1000, (det_time - solve_time) * 1000, (
+          inverse_time - det_time) * 1000, (inverse_time - start_time) * 1000))
+
+
+if __name__ == '__main__':
+  main()
