+++ python/constants.py
@@ -1,0 +1,151 @@
+A1_21 = [
+    [26, 65, -32, -35, 91, 70, 27, -96, 25],
+    [-55, 88, -11, 71, -15, -18, -10, 29, -46],
+    [-69, 12, -78, 52, -93, -77, -95, 3, -20],
+    [-86, -11, -60, -83, -1, -39, 54, 13, 41],
+    [-61, -22, 99, -56, -64, -79, -46, 53, -58],
+    [41, -46, -18, 84, 69, 38, -71, -84, -26],
+    [87, -14, -60, 40, 12, 13, -58, -18, -50],
+    [93, -91, 65, -85, -26, -12, 91, 4, 58],
+    [33, -34, -75, -72, -66, 15, 84, 11, -72],
+]
+
+b1_21 = [
+    [-69],
+    [1],
+    [-45],
+    [-79],
+    [-97],
+    [-72],
+    [87],
+    [44],
+    [-15],
+]
+
+A2_21 = [
+    [188, -32, -9, 7, -61, 49, -53, 23, -31, 62],
+    [-32, 101, 11, -32, -80, 1, 53, -7, 94, -109],
+    [-9, 11, 231, 96, -62, -12, 72, -40, 72, -77],
+    [7, -32, 96, 174, -37, -119, 19, 7, -27, 27],
+    [-61, -80, -62, -37, 184, -34, -115, 45, -53, 46],
+    [49, 1, -12, -119, -34, 182, 80, -56, -4, 18],
+    [-53, 53, 72, 19, -115, 80, 231, -101, 70, -50],
+    [23, -7, -40, 7, 45, -56, -101, 227, -10, 23],
+    [-31, 94, 72, -27, -53, -4, 70, -10, 241, -155],
+    [62, -109, -77, 27, 46, 18, -50, 23, -155, 212],
+]
+
+b2_21 = [
+    [-44],
+    [74],
+    [87],
+    [247],
+    [-39],
+    [168],
+    [192],
+    [-199],
+    [-194],
+    [243],
+]
+
+A3_21 = [
+    [390, 201, 0, 0, 0, 0, 0, 0, 0, 0],
+    [21, -135, -53, 0, 0, 0, 0, 0, 0, 0],
+    [0, -271, 868, -203, 0, 0, 0, 0, 0, 0],
+    [0, 0, -7, 596, 160, 0, 0, 0, 0, 0],
+    [0, 0, 0, -47, -208, -96, 0, 0, 0, 0],
+    [0, 0, 0, 0, -24, 347, -17, 0, 0, 0],
+    [0, 0, 0, 0, 0, -96, 220, -40, 0, 0],
+    [0, 0, 0, 0, 0, 0, 107, 546, -204, 0],
+    [0, 0, 0, 0, 0, 0, 0, -301, -740, -364],
+    [0, 0, 0, 0, 0, 0, 0, 0, 48, -154],
+]
+
+b3_21 = [
+    [446],
+    [62],
+    [-783],
+    [264],
+    [-747],
+    [-732],
+    [-803],
+    [-716],
+    [-664],
+    [-608],
+]
+
+A1_30 = [
+    [-27, -93, -74, -45, 67, 74, 17, 96, -36],
+    [59, 79, -66, 45, -90, 20, -58, -61, -62],
+    [73, -82, -1, 9, -79, -27, 40, 26, 52],
+    [-39, 91, -95, 80, 52, -26, -14, -10, -42],
+    [-23, -64, -10, 61, -33, -98, -28, 12, 22],
+    [17, 45, 23, 58, 27, -27, -78, 99, -64],
+    [-39, 27, 6, -66, 87, 32, -15, -36, 82],
+    [89, 40, -17, -44, -29, -31, 55, -52, 3],
+    [81, -60, 86, 42, 43, 29, 95, -70, -26],
+]
+
+B1_30 = [
+    [43],
+    [84],
+    [-7],
+    [38],
+    [59],
+    [93],
+    [-5],
+    [73],
+    [-10],
+]
+
+A2_30 = [
+    [688, -245, -382, 56, 166, -290, -5, -49, -347, -279],
+    [-245, 652, 382, 235, -226, -28, -73, -132, -148, 96],
+    [-382, 382, 705, 96, -41, 43, -30, 198, 182, 193],
+    [56, 235, 96, 611, 43, 280, -289 - 158, 74, 179],
+    [166, -226, -41, 43, 166, -4, -51, 143, 77, 66],
+    [-290, -28, 43, 280, -4, 677, 107, -151, 395, 183],
+    [-5, -73, -30, -289, -51, 107, 691, -158, 128, -74],
+    [-49, -132, 198, -158, 143, -151, -158, 498, 36, 223],
+    [-347, -148, 182, 74, 77, 395, 128, 36, 668, 299],
+    [-279, 96, 193, 179, 66, 183, -74, 223, 299, 751],
+]
+
+B2_30 = [
+    [-324],
+    [79],
+    [671],
+    [553],
+    [-120],
+    [-68],
+    [-173],
+    [15],
+    [644],
+    [22]
+]
+
+A3_30 = [
+    [949, 134, 0, 0, 0, 0, 0, 0, 0, 0],
+    [53, 994, 15, 0, 0, 0, 0, 0, 0, 0],
+    [0, -24, -339, -3, 0, 0, 0, 0, 0, 0],
+    [0, 0, 332 - 858, -373, 0, 0, 0, 0, 0],
+    [0, 0, 0, 41, -128, -13, 0, 0, 0, 0],
+    [0, 0, 0, 0, 72, 227, 88, 0, 0, 0],
+    [0, 0, 0, 0, 0, -267, 863, -208, 0, 0],
+    [0, 0, 0, 0, 0, 0, 74, 796, 3, 0],
+    [0, 0, 0, 0, 0, 0, 0, 72, 226, 7],
+    [0, 0, 0, 0, 0, 0, 0, 0, -55, -596],
+]
+
+B3_30 = [
+    [-144],
+    [933],
+    [240],
+    [391],
+    [441],
+    [-306],
+    [34],
+    [113],
+    [-687],
+    [124],
+]
